+++ src/assets/env.template.js
@@ -28,6 +28,9 @@
   window['env']['matomoSiteId'] = '$MATOMO_SITE_ID';
   window['env']['matomoSiteUrl'] = '$MATOMO_SITE_URL';
 
+  //Sentry integration
+  window['env']['sentryDsn'] = '$SENTRY_DSN';
+
   // Loan submission button disabled timeout in seconds
   window['env']['loanSubmitButtonDisabledTimeOut'] = '$LOAN_SUBMIT_DISABLED_TIME_OUT';
 })(this);