+++ src/environments/environment.ts
@@ -37,6 +37,8 @@ export const environment = {
   matomoSiteUrl: window['env']['matomoSiteUrl'] || 'https://analytics.integration.oneacrefund.org',
   // Loan submission button disabled timeout in seconds
   loanSubmitButtonDisabledTimeOut: window['env']['loanSubmitButtonDisabledTimeOut'] || 5,
+  sentryDsn: window['env']['sentryDsn'] || 'https://test@o454511.ingest.us.sentry.io/4508318875844608'
+
 };
 
 // Server URL